webui: cleanup Debian PHP dependencies
Debian PHP dependencies now uses the platform default, but offers the possibility to use PHP5, if available.
The previous used dependencies php7 have been incorrect. It should have been php7.0.
However, in the next Debian release it will be php7.2, so sticking to the platform default might be the best option.
Apache PHP is not enabled any longer,
as this is done by the package libapache2-mod-php*.